moves "card exists" to common card step defs
