+++ src/test/java/com/codurance/retropolis/acceptance/cards/UpdateCardStepDefinitionIntegrationTest.java
@@ -1,14 +1,12 @@
 package com.codurance.retropolis.acceptance.cards;
 
 import static com.codurance.retropolis.utils.HttpWrapper.executePatch;
-import static com.codurance.retropolis.utils.HttpWrapper.executePost;
 import static com.codurance.retropolis.utils.HttpWrapper.responseResult;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
 import com.codurance.retropolis.acceptance.BaseStepDefinition;
 import com.codurance.retropolis.entities.Card;
-import com.codurance.retropolis.requests.NewCardRequestObject;
 import com.codurance.retropolis.requests.UpdateCardRequestObject;
 import com.codurance.retropolis.responses.CardResponseObject;
 import com.codurance.retropolis.utils.HttpWrapper;
@@ -16,7 +14,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.cucumber.java.en.And;
-import io.cucumber.java.en.Given;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import javax.sql.DataSource;
@@ -25,18 +22,8 @@ import org.springframework.http.HttpEntity;
 
 public class UpdateCardStepDefinitionIntegrationTest extends BaseStepDefinition {
 
-  private final String CARD_TEXT = "Hello";
-  private final Long COLUMN_ID = 1L;
-  private final String USER_EMAIL = "john.doe@codurance.com";
-
   public UpdateCardStepDefinitionIntegrationTest(DataSource dataSource) {
     super(dataSource);
-  }
-
-  @Given("the card exists")
-  public void theCardExistsWithId() {
-    executePost(url + "/cards",
-        new HttpEntity<>(new NewCardRequestObject(CARD_TEXT, COLUMN_ID, USER_EMAIL), headers));
   }
 
   @When("the client updates to cards with this id and changes the text to {string}")
